Move featureflag to api
