+++ src/api/modules/index.ts
@@ -6,3 +6,4 @@ export * from './sponsor'
 export * from './workspace'
 export * from './auth'
 export { default as authUrl } from './url'
+export * from './featureFlag'
